Sentiment Analysis of Video Game Reviews
MSc Project (The University of Kolkata, 2018): The study analyzed the reviews based on a dictionary of positive and negative entities.  Classification Algorithms based on the data were trained, which in turn were used to reclassify the same content to see how well each algorithm performed.  The goal was to compare the efficacy of different classifying methods in categorizing text by semantic meaning.