Statistical Learning Project
In this group project, we address the problem of classifying celestial objects using light spectrum readings as predictors.  The focus was to accurately classify the objects with as low an out-of-sample error rate as possible.  An overview of certain key astrophysics concepts and a description of the main statistical learning models attempted were provided.